models.py -- changed Assignment so that it contains a BooleanField indicating whether or not multiple submissions are allowed.
views.py -- edited submit_assignment so that it gives user a "cannot submit"
page if they try to resubmit for a single-submission-only assignment.
Should probably remove submission link or grey out button or something.

cannot_submit.html -- a simple template which just displays an "error message",
which is either that submissions are closed or that the user has already submitted.